face identifing system/requirements.txt
streamlit
numpy
joblib
tensorflow
pillow
